Move backend chat tests into subdirectories
